backend/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
pymongo==4.6.1
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1

# AI Services
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
mistralai==0.0.8

# Utilities
requests==2.31.0
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
marshmallow==3.20.1